+++ packages/@headlessui-react/src/components/combobox/combobox.tsx
@@ -43,6 +43,7 @@ import { useOpenClosed, State, OpenClosedProvider } from '../../internal/open-cl
 import { Keys } from '../keyboard'
 import { useControllable } from '../../hooks/use-controllable'
 import { useWatch } from '../../hooks/use-watch'
+import { useTrackedPointer } from '../../hooks/use-tracked-pointer'
 
 enum ComboboxState {
   Open,
@@ -1255,13 +1256,19 @@ let Option = forwardRefWithAs(function Option<
     actions.goToOption(Focus.Specific, id)
   })
 
-  let handleMove = useEvent(() => {
+  let pointer = useTrackedPointer()
+
+  let handleEnter = useEvent((evt) => pointer.update(evt))
+
+  let handleMove = useEvent((evt) => {
+    if (!pointer.wasMoved(evt)) return
     if (disabled) return
     if (active) return
     actions.goToOption(Focus.Specific, id, ActivationTrigger.Pointer)
   })
 
-  let handleLeave = useEvent(() => {
+  let handleLeave = useEvent((evt) => {
+    if (!pointer.wasMoved(evt)) return
     if (disabled) return
     if (!active) return
     if (data.optionsPropsRef.current.hold) return
@@ -1286,6 +1293,8 @@ let Option = forwardRefWithAs(function Option<
     disabled: undefined, // Never forward the `disabled` prop
     onClick: handleClick,
     onFocus: handleFocus,
+    onPointerEnter: handleEnter,
+    onMouseEnter: handleEnter,
     onPointerMove: handleMove,
     onMouseMove: handleMove,
     onPointerLeave: handleLeave,