+++ packages/@headlessui-react/src/test-utils/interactions.ts
@@ -1,5 +1,6 @@
 import { fireEvent } from '@testing-library/react'
 import { disposables } from '../utils/disposables'
+import { pointer } from './fake-pointer'
 
 let d = disposables()
 
@@ -318,8 +319,13 @@ export async function mouseMove(element: Document | Element | Window | null) {
   try {
     if (element === null) return expect(element).not.toBe(null)
 
-    fireEvent.pointerMove(element)
-    fireEvent.mouseMove(element)
+    pointer.advance()
+
+    pointer.bypassingTrackingChecks(() => {
+      fireEvent.pointerMove(element)
+    })
+
+    fireEvent.mouseMove(element, pointer.options)
 
     await new Promise(nextFrame)
   } catch (err) {
@@ -332,10 +338,15 @@ export async function mouseLeave(element: Document | Element | Window | null) {
   try {
     if (element === null) return expect(element).not.toBe(null)
 
-    fireEvent.pointerOut(element)
-    fireEvent.pointerLeave(element)
-    fireEvent.mouseOut(element)
-    fireEvent.mouseLeave(element)
+    pointer.advance()
+
+    pointer.bypassingTrackingChecks(() => {
+      fireEvent.pointerOut(element)
+      fireEvent.pointerLeave(element)
+    })
+
+    fireEvent.mouseOut(element, pointer.options)
+    fireEvent.mouseLeave(element, pointer.options)
 
     await new Promise(nextFrame)
   } catch (err) {