+++ packages/@headlessui-vue/src/components/combobox/combobox.ts
@@ -35,6 +35,7 @@ import { useOutsideClick } from '../../hooks/use-outside-click'
 import { Hidden, Features as HiddenFeatures } from '../../internal/hidden'
 import { objectToFormEntries } from '../../utils/form'
 import { useControllable } from '../../hooks/use-controllable'
+import { useTrackedPointer } from '../../hooks/use-tracked-pointer'
 
 function defaultComparator<T>(a: T, z: T): boolean {
   return a === z
@@ -1057,13 +1058,21 @@ export let ComboboxOption = defineComponent({
       api.goToOption(Focus.Specific, id)
     }
 
-    function handleMove() {
+    let pointer = useTrackedPointer()
+
+    function handleEnter(evt: PointerEvent) {
+      pointer.update(evt)
+    }
+
+    function handleMove(evt: PointerEvent) {
+      if (!pointer.wasMoved(evt)) return
       if (props.disabled) return
       if (active.value) return
       api.goToOption(Focus.Specific, id, ActivationTrigger.Pointer)
     }
 
-    function handleLeave() {
+    function handleLeave(evt: PointerEvent) {
+      if (!pointer.wasMoved(evt)) return
       if (props.disabled) return
       if (!active.value) return
       if (api.optionsPropsRef.value.hold) return
@@ -1086,6 +1095,8 @@ export let ComboboxOption = defineComponent({
         disabled: undefined, // Never forward the `disabled` prop
         onClick: handleClick,
         onFocus: handleFocus,
+        onPointerenter: handleEnter,
+        onMouseenter: handleEnter,
         onPointermove: handleMove,
         onMousemove: handleMove,
         onPointerleave: handleLeave,